% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_1.R
\docType{data}
\name{dataset_1}
\alias{dataset_1}
\title{Synthetic dataset 1}
\format{
`dataset_1`
A data frame with 1,000 rows and 2 columns:
\describe{
  \item{X1}{Continuous variable, mean = 1, sd = 1; mean = 8, sd = 1}
  \item{X2}{Continuous variable, mean = 2, sd = 1; mean = 9, sd = 1}
}
}
\source{
Synthetic
}
\usage{
dataset_1
}
\description{
A synthetic dataset comprised of two distinct clusters (1:500, 501:1000), where the values of both predictors do not overlap each other. One variable is informative of the other variable.
}
\keyword{datasets}
