% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_3.R
\docType{data}
\name{dataset_3}
\alias{dataset_3}
\title{Synthetic dataset 3}
\format{
## `dataset_3`
A data frame with 1,000 rows and 3 columns:
\describe{
  \item{X1}{Categorical variable, levels = A or B, distinct for each cluster}
  \item{X2}{Categorical variable, levels = 1 or 2, distinct for each cluster}
}
}
\source{
Synthetic
}
\usage{
dataset_3
}
\description{
A synthetic dataset comprised of two clusters (1:500, 501:1000), where X1 is distinct between both clusters and X2 is distinct between both clusters. Both variables are informative of the cluster.
}
\keyword{datasets}
