% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_4.R
\docType{data}
\name{dataset_4}
\alias{dataset_4}
\title{Synthetic dataset 4}
\format{
## `dataset_4`
A data frame with 500 rows and 5 columns:
\describe{
  \item{X1}{Continuous variable, mean = 8, sd = 3}
  \item{X2}{Continuous variable, mean = 2, sd = 3}
  \item{X3}{Continuous variable, mean = -1, sd = 3}
  \item{X4}{Categorical variable, levels = 0 or 1}
  \item{Y}{Continuous variable, response of regression}
}
}
\source{
Synthetic
}
\usage{
dataset_4
}
\description{
A synthetic dataset, where Y is the outcome variable, X1/X2/X3 are continuous predictors and X4 is a categorical predictor.
}
\keyword{datasets}
