% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_6.R
\docType{data}
\name{dataset_6}
\alias{dataset_6}
\title{Synthetic dataset 6}
\format{
## `dataset_6`
A data frame with 1,000 rows and 2 columns:
\describe{
  \item{X1}{Continuous variable, mean = 1, sd = 1; mean = 8, sd = 1}
  \item{X2}{Continuous variable, mean = 2, sd = 1; mean = 9, sd = 1}
}
}
\source{
Synthetic
}
\usage{
dataset_6
}
\description{
A synthetic dataset comprised of two distinct clusters (1:500, 501:1000), where the values of both predictors do not overlap each other. One variable is informative of the other variable. The clusters are defined so that they have a specific shape.
}
\keyword{datasets}
