% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Diagnostic Plot}
\usage{
plot_alpha(x, nburn, thinning, ...)
}
\arguments{
\item{x}{object class 'dpmm_fit'.}

\item{nburn}{Iterations to be discarded in the plot of component rankings}

\item{thinning}{Iterations}

\item{...}{other parameters used by ggplot2.}
}
\value{
A 'ggplot' panel plot.
}
\description{
Number of Components, Component ranking and trace plot of alpha values.
}
\examples{
\dontrun{
## load dataset
data(dataset_1)

## fit model
posteriors <- runModel(dataset_1, 
                       mcmc_iterations = 100,
                       L = 6, 
                       mcmc_chains = 2, 
                       standardise = TRUE)
               
                              
## diagnostic plot
plot_alpha(posteriors, nburn = 50, thinning = 1)
}


}
