% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dpmm_fit.R
\name{plot_dpmm_fit}
\alias{plot_dpmm_fit}
\title{Plot Parameter Trace plots}
\usage{
plot_dpmm_fit(x, trace = TRUE, density = TRUE, ...)
}
\arguments{
\item{x}{object class 'dpmm_fit'}

\item{trace}{Logical parameter for plotting trace plots. (default = TRUE)}

\item{density}{Logical parameter for plotting density plots. (default = TRUE)}

\item{...}{other arguments used by base plot.}
}
\value{
Plot
}
\description{
This function plots trace plots of the DPMM estimated parameters.
}
\examples{
\dontrun{
## load dataset
data(dataset_1)

## fit model
posteriors <- runModel(dataset_1, 
                       mcmc_iterations = 100,
                       L = 6, 
                       mcmc_chains = 2, 
                       standardise = TRUE)
## plot trace plots
plot(posteriors)
}


}
