% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dpmm_fit.R
\name{predict_dpmm_fit}
\alias{predict_dpmm_fit}
\title{Predict from DPMM}
\usage{
predict_dpmm_fit(object, newdata, samples = seq(1000, 2500, 100), ...)
}
\arguments{
\item{object}{object class 'dpmm_fit'}

\item{newdata}{dataframe with missingness}

\item{samples}{vector of iterations to be used in the posterior}

\item{...}{other parameters used in 'posterior.dpmm'}
}
\value{
A list with n entries for n rows with missingness, each entry is a dataframe with the sampled missing values.
}
\description{
Predict missing values from the fitted DPMM.
}
\examples{
## load dataset
data(dataset_1)

## fit model
posteriors <- runModel(dataset_1, 
                       mcmc_iterations = 100,
                       L = 6, 
                       mcmc_chains = 2, 
                       standardise = TRUE)
                       
## introduce missing data
rows <- 501:550
dataset_missing <- dataset_1
dataset_missing_predict <- dataset_missing[rows,]
dataset_missing_predict[,1] <- as.numeric(NA)

# predict missing values
posteriors.dpmmfit <- predict(posteriors, 
                              dataset_missing_predict, 
                              samples = c(1:100))
                       

}
