% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Fit DPMM}
\usage{
runModel(
  dataset,
  mcmc_iterations = 2500,
  L = 10,
  mcmc_chains = 2,
  standardise = TRUE
)
}
\arguments{
\item{dataset}{Dataframe with continuous and/or categorical variables.}

\item{mcmc_iterations}{Number of MCMC iterations.}

\item{L}{Number of DPMM components fitted.}

\item{mcmc_chains}{Number of MCMC chains fitted.}

\item{standardise}{Should continuous variables be standardised. default = TRUE}
}
\value{
A list.
}
\description{
This function fits the Dirichlet Process Mixture Model (DPMM).
}
\examples{
## load dataset
data(dataset_1)

## fit model
posteriors <- runModel(dataset_1, mcmc_iterations = 100, L = 6, mcmc_chains = 2, standardise = TRUE)

}
