% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_conditional_RW.R
\name{sampler_conditional_RW}
\alias{sampler_conditional_RW}
\title{Conditional Random Walk Sampler}
\usage{
sampler_conditional_RW(model, mvSaved, target, control)
}
\description{
Nimble Sampler for a missing continuous predictor with other observed predictors.
}
