% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_conditional_RW_block.R
\name{sampler_conditional_RW_block}
\alias{sampler_conditional_RW_block}
\title{Conditional Block Random Walk Sampler}
\usage{
sampler_conditional_RW_block(model, mvSaved, target, control)
}
\description{
Nimble Sampler for >1 missing continuous predictors with other observed predictors.
}
