% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLogPostDens.R
\name{sumLogPostDens}
\alias{sumLogPostDens}
\title{Sum-posterior-log-density}
\usage{
sumLogPostDens(model, mvSaved, target, control)
}
\arguments{
\item{model}{.}

\item{mvSaved}{.}

\item{target}{.}

\item{control}{.}
}
\description{
Nimble Sampler function to record the sum of the (unnormalized) model posterior log-density, on every iteration of an MCMC. This is based on the vignette https://danielturek.github.io/public/sumPostLogDens/sumPostLogDens.html
}
